+++ __init__.py
@@ -1,9 +1,10 @@
-from .Core.base import Vec1, Vec2, Color3, Color4, c256, Key, MouseButton, drawMode, shaderType, textureType, stretchType
+from .Core.base import Vec1, Vec2, Color3, Color4, c256, Key, MouseButton, drawMode, shaderType, textureType, stretchType, batchDrawing
 from .Graphics.window.window import Window
-from .Graphics.window.gui import SimpleButton
+from .Graphics.window.gui import SimpleButton, Frame, textInput
 from .Control.keyboard import Keyboard
 from .Control.mouse import Mouse
 from .Graphics import *
+from .Graphics.render.batchRender import batchRender
 from .Utils.General.texture import loadTexture
 from .Utils.General.frametimer import frameTimer
 from .Utils.General.saver import saveData, loadData