+++ CHANGELOG.md
@@ -11,7 +11,12 @@ added **frameTimer** + **Particle system** + **Update README.md** + **Improved w
 updated **Simple particles**, **log_system**
 added **randomColor3** + **randomColor4** + **genKey** + **saveData** + **loadData**
 
-# 25.06.2026
+# 25.06.26
 
 updated **utils structure**
 added **checkInDebbuger** + **requirements.txt** + **builtin icons** + **loadSound** + **soundManager** + **sceneManager**
+
+# 26.06.26
+
+updated **Window** + **README.md**
+added **batchRender** + **Frame** + **textInput** + **memoryClean** (Clean memory before exit)